+++ allo/backend/__init__.py
@@ -4,7 +4,7 @@ import os
 from . import llvm, hls, ip
 
 try:
-    from . import experimental
-    from .experimental import AIE_MLIRModule
+    from . import aie
+    from .aie import AIE_MLIRModule
 except ImportError:
     AIE_MLIRModule = None
